+++ packages/profiler/lib/sdk.js
@@ -1,0 +1,6 @@
+import { Profiler } from "./hantryProfiler";
+
+export const init = (dsn, options) => {
+  const profiler = new Profiler(dsn, options);
+  profiler.observerStart();
+};
